// src/components/Tag/Tag.types.ts
// Generated with util/create-component.js
import {Intent} from '../../common/';
export interface TagProps {
    fill?: boolean;
    // icon?: boolean;
    intent?: Intent;
    interactive?: boolean;
    large?: boolean;
    minimal?: boolean;
    // removable?: boolean;
    // rightIcon?: boolean;
    round?: boolean;
    // tags?: string[];
}